fixed reset comment area on mobile
